Farvel /flavors/detail. Also consolidating the API() into a common class
